malicious_content_search/requirements.txt
asgiref==3.7.2
asyncpg==0.28.0
autopep8==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.0.0
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-quill-editor==0.1.40
django-requests==0.0.2
django-session-timeout==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
pycodestyle==2.11.0
pyotp==2.9.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
Unipath==1.1
urllib3==2.0.4
whitenoise==6.5.0
